Add epoll and do more release readiness changes
This change also pays off some of the remaining technical debt with
stdio, file descriptors, and memory managemnt polyfills.

+++ tool/build/lib/syscall.c
@@ -1227,7 +1227,11 @@ static int OpGetPpid(struct Machine *m) {
 }
 
 static int OpKill(struct Machine *m, int pid, int sig) {
-  return kill(pid, sig);
+  if (pid == getpid()) {
+    ThrowProtectionFault(m);
+  } else {
+    return kill(pid, sig);
+  }
 }
 
 static int OpGetUid(struct Machine *m) {
